python/requirements.txt
Pillow==9.4.0
PyOpenGL==3.1.6
numpy==1.24.2